magento-issue-7852: Moved the review functionality from Magento_Customer module to Magento_Review module for removing dependency of Customer module on Review module.
