+++ app/code/Magento/Review/Controller/Adminhtml/Customer/ProductReviews.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Customer\Controller\Adminhtml\Index;
+namespace Magento\Review\Controller\Adminhtml\Customer;
 
 class ProductReviews extends \Magento\Customer\Controller\Adminhtml\Index
 {
